Make wired network settings functional
Fetch all network devices and show local IP. This removes public IP from
display but leaves the code to fetch it for possible future use.

Note that this does not attempt to filter out wireless or virtual
network devices so those may be shown on Wired page now.

Bug-AGL: SPEC-5296
Change-Id: I52de896ed3827c86782ee1cf29208a295532c3f2
Signed-off-by: Tomi Leppänen <tomi.leppanen@seafarix.com>